requirements.txt
nltk.download('stopwords')
nltk.download('averaged_perceptron_tagger')
nltk.download('wordnet')
sklearn
vaderSentiment
pandas
tweepy
